requirements.txt
requests
beautifulsoup4
streamlit
psycopg2-binary
python-dotenv